FEATURE: Implemented sharing of lists. We can now share our lists with any number of users in the system. We can also assign read, write and delete rights to the user. All api's check access rights before they allow an operation to happen. This does add a little extra overhead per api call but I've tried to keep it to a minimum.

+++ model/shoplisttable.php
@@ -1,5 +1,11 @@
 <?php
 require_once dirname(__FILE__) . DIRECTORY_SEPARATOR . "tablebase.php";
+require_once dirname(__FILE__) . DIRECTORY_SEPARATOR . "accesscontrolledobject.php";
+require_once dirname(__FILE__) . DIRECTORY_SEPARATOR . "shoplist.php";
+require_once dirname(__FILE__) . DIRECTORY_SEPARATOR . "permissions.php";
+
+
+if (!class_exists('ListFunctorShopList')) {
 
 class ListFunctorShopList
 {
@@ -10,17 +16,15 @@ class ListFunctorShopList
         $this->_shoplist = & $shoplist_array;
     }
 
-    public function iterate_row($list_id, $list_name, $item_count)
+    public function iterate_row(ShoppingList $shop_list)
     {
-        $thisItem = array(
-            ShopListTable::kCol_ListID => $list_id,
-            ShopListTable::kCol_ListName => $list_name,
-            ShopListTable::kCol_ItemCount => $item_count);
-
-        $this->_shoplist[] = $thisItem;
+        $this->_shoplist[] = $shop_list;
     }
 }
+}
 
+if (!class_exists('ShopListTable')) {
+	
 class ShopListTable extends TableBase
 {
 	const kTableName = 'shoplists';
@@ -41,20 +45,46 @@ class ShopListTable extends TableBase
 	{
 		if ($fetch_count > 0) {
 				
-			$sql = sprintf("SELECT sl.listID, sl.listName, (
-								SELECT COUNT(`listID_FK`) 
-								FROM `shopitems` si 
-								where si.`listID_FK`=sl.`listID` AND si.`isPurchased` <= 0) AS itemCount
-							FROM shoplists sl
-							WHERE sl.`userID_FK`=%d",
+			$sql = sprintf("SELECT 
+								sl.listID, 
+								sl.listName,
+								sl.userID_FK, 
+								(SELECT 
+									COUNT(`listID_FK`) 
+								FROM 
+									`shopitems` si 
+								WHERE 
+									si.`listID_FK`=sl.`listID` AND si.`isPurchased` <= 0) AS itemCount,
+								sls.user_perms,
+								sls.user_id_FK
+							FROM 
+								shoplists sl
+							LEFT JOIN 
+								shoplists_sharing sls
+							ON 
+								sl.`listID`=sls.`list_id_FK`
+							WHERE 
+								sl.`userID_FK`=%d OR sls.`user_id_FK`=%d",
+					parent::GetUserID(),
 					parent::GetUserID());
  
 		} else {
 				
-			$sql = sprintf("
-					SELECT shoplists.listID, shoplists.listName
-					FROM shoplists
-					WHERE shoplists.userID_FK=%d",
+			$sql = sprintf("SELECT 
+								sl.listID, 
+								sl.listName, 
+								sl.userID_FK,
+								sls.user_perms,
+								sls.user_id_FK
+							FROM 
+								shoplists sl
+							LEFT JOIN 
+								shoplists_sharing sls
+							ON 
+								sl.`listID`=sls.`list_id_FK`
+							WHERE 
+								sl.userID_FK=%d OR sls.`user_id_FK`=%d",
+					parent::GetUserID(),
 					parent::GetUserID());
 		}	
 		
@@ -63,25 +93,30 @@ class ShopListTable extends TableBase
 			throw new Exception("SQL exec failed (" . $this->get_db_con()->errno . ")");
 		}
 		
-		while ($row = mysqli_fetch_array($result)) {
-			
-			if ($fetch_count > 0) {
+		while ($row = mysqli_fetch_array($result, MYSQL_ASSOC)) {
+
+			$thisList = ShoppingList::create_from_fields(
+					intval($row['listID']),
+					$row['listName'],
+					intval($row['userID_FK']),
+					($fetch_count > 0) ? intval($row['itemCount']) : 0,
+					!is_null($row['user_id_FK']) ? intval($row['user_id_FK']) : 0,
+					!is_null($row['user_perms']) ? intval($row['user_perms']) : 0);
+
+			if (Permissions::can_read(parent::GetUserID(), $thisList)) {
 				call_user_func(
 					array($functor_obj, $function_name), // invoke the callback function
-					$row[0], // 'listID' 
-					$row[1], // 'listName'		
-					$row[2]);// 'itemCount'
-			} else {
-				call_user_func(
-					array($functor_obj, $function_name), // invoke the callback function
-					$row[0], // 'listID' 
-					$row[1], // 'listName'		
-					0);// 'itemCount'
+					$thisList);
 			}
 		}
 		
 		if ($result)
 			$result->free();
+	}
+	
+	function get_list($list_id) {
+		
+		return ShoppingList::create_from_db($this->get_db_con(), $list_id, parent::GetUserID());		
 	}
 	
 	function add_list($list_name)
@@ -112,42 +147,59 @@ class ShopListTable extends TableBase
 		if (!$config['USE_STORED_PROCS']) {
 
 			try {
-				$isTransactional = $this->get_db_con()->autocommit(FALSE);
-				if ($isTransactional == FALSE) {
-					throw new Exception("Could Not Start Transaction.");
-				}
+				$this_list = $this->get_list($list_ID);
 				
-				$sql = sprintf("DELETE FROM `shopitems` 
-								WHERE `%s`=%d AND `listID_FK`=%d",
-								self::kCol_UserID,
-								parent::GetUserID(),
-								$list_ID);
-			
-				$result = $this->get_db_con()->query($sql);
-				if ($result == FALSE) {
-					throw new Exception(
-						"Error deleting Shopping List (" . 
-						$this->get_db_con()->errno . ")");
-				}
+				if (Permissions::can_delete(parent::GetUserID(), $this_list)) {				
+					
+					$isTransactional = $this->get_db_con()->autocommit(FALSE);
+					if ($isTransactional == FALSE) {
+						throw new Exception("Could Not Start Transaction.");
+					}
+					
+					$sql = sprintf("DELETE FROM `shopitems` 
+									WHERE `" . self::kCol_UserID . "`=%d AND `listID_FK`=%d",
+									$this_list->getOwnerId(),
+									$list_ID);
 				
-				$sql = sprintf(
-					"DELETE FROM `%s` WHERE `%s`=%d AND `%s`=%d",
-					self::kTableName,
-					self::kCol_ListID,
-					$list_ID,
-					self::kCol_UserID,
-					parent::GetUserID());
-	
-				$result = $this->get_db_con()->query($sql);
-				if ($result == FALSE) {
-					throw new Exception(
-						"Error deleting Shopping List (",
-						$this->get_db_con()->errno . ")");
+					$result = $this->get_db_con()->query($sql);
+					if ($result == FALSE) {
+						throw new Exception(
+							"Error deleting Shopping List (" . 
+							$this->get_db_con()->errno . ")");
+					}
+
+					$sql = sprintf("DELETE FROM shoplists_sharing
+									WHERE list_ID_FK=%d",
+									$list_ID);
+					
+					$result = $this->get_db_con()->query($sql);
+					if ($result == FALSE) {
+						throw new Exception(
+								"Error deleting Shopping List (" .
+								$this->get_db_con()->errno . ")");
+					}
+						
+					$sql = sprintf(
+						"DELETE FROM `%s` WHERE `%s`=%d AND `%s`=%d",
+						self::kTableName,
+						self::kCol_ListID,
+						$list_ID,
+						self::kCol_UserID,
+						parent::GetUserID());
+		
+					$result = $this->get_db_con()->query($sql);
+					if ($result == FALSE) {
+						throw new Exception(
+							"Error deleting Shopping List (",
+							$this->get_db_con()->errno . ")");
+					}
+					
+					$this->get_db_con()->commit();
+					$this->get_db_con()->autocommit(TRUE);
+					
+				} else {
+					throw new Exception("You don't have permissions to perform this operation.");
 				}
-				
-				$this->get_db_con()->commit();
-				$this->get_db_con()->autocommit(TRUE);
-			
 			} catch (Exception $e) {
 					
 				if ($isTransactional) {
@@ -176,17 +228,25 @@ class ShopListTable extends TableBase
 	
 	function edit_list($list_ID, $list_name)
 	{
-		$sql = sprintf(
-				"UPDATE %s SET `listName`='%s' WHERE `listID`=%d AND `userID_FK`=%d", 
-				self::kTableName,
-				$this->get_db_con()->escape_string($list_name),
-				$list_ID,
-				$this->GetUserID());
+		$this_list = $this->get_list($list_ID);
 		
-		$result = $this->get_db_con()->query($sql);
-		if ($result == FALSE) {
-			throw new Exception ("Failed to rename list: (" . $this->get_db_con()->errno);
+		if (Permissions::can_write(self::GetUserID(), $this_list)) {
+			
+			$sql = sprintf(
+					"UPDATE %s SET `listName`='%s' WHERE `listID`=%d AND `userID_FK`=%d", 
+					self::kTableName,
+					$this->get_db_con()->escape_string($list_name),
+					$list_ID,
+					$this->GetUserID());
+			
+			$result = $this->get_db_con()->query($sql);
+			if ($result == FALSE) {
+				throw new Exception ("Failed to rename list: (" . $this->get_db_con()->errno);
+			}
+		} else {
+			throw new Exception("You don't have permissions to perform this operation.");
 		}
 	}
+}
 }
 ?>
